Corrected the main crystal in marineb.cpp to 12MHz as seen in PCB photos.
Cleaned up clock definitions in galaxian driver.

Moved ladybugg from galaxold to galaxian drivers.
