+++ src/mame/drivers/1942.cpp
@@ -62,13 +62,15 @@ correctly.
 ***************************************************************************/
 
 /* 12mhz OSC */
-#define MAIN_CPU_CLOCK      (XTAL_12MHz/3)
-#define SOUND_CPU_CLOCK     (XTAL_12MHz/4)
-#define AUDIO_CLOCK     (XTAL_12MHz/8)
+#define MASTER_CLOCK      (XTAL_12MHz)
+#define MAIN_CPU_CLOCK      (MASTER_CLOCK/3)
+#define SOUND_CPU_CLOCK     (MASTER_CLOCK/4)
+#define AUDIO_CLOCK     (MASTER_CLOCK/8)
 /* 20mhz OSC - both Z80s are 4 MHz */
-#define MAIN_CPU_CLOCK_1942P      (XTAL_20MHz/5)
-#define SOUND_CPU_CLOCK_1942P     (XTAL_20MHz/5)
-#define AUDIO_CLOCK_1942P     (XTAL_20MHz/16)
+#define MASTER_CLOCK_1942P     (XTAL_20MHz)
+#define MAIN_CPU_CLOCK_1942P      (MASTER_CLOCK_1942P/5)
+#define SOUND_CPU_CLOCK_1942P     (MASTER_CLOCK_1942P/5)
+#define AUDIO_CLOCK_1942P     (MASTER_CLOCK_1942P/16)
 
 #include "emu.h"
 #include "cpu/z80/z80.h"
