+++ src/mame/drivers/marineb.cpp
@@ -534,7 +534,7 @@ INTERRUPT_GEN_MEMBER(marineb_state::wanted_vblank_irq)
 static MACHINE_CONFIG_START( marineb, marineb_state )
 
 	/* basic machine hardware */
-	MCFG_CPU_ADD("maincpu", Z80, 3072000)   /* 3.072 MHz */
+	MCFG_CPU_ADD("maincpu", Z80, CPU_CLOCK)   /* 3 MHz? */
 	MCFG_CPU_PROGRAM_MAP(marineb_map)
 	MCFG_CPU_IO_MAP(marineb_io_map)
 	MCFG_CPU_VBLANK_INT_DRIVER("screen", marineb_state,  marineb_vblank_irq)
@@ -555,7 +555,7 @@ static MACHINE_CONFIG_START( marineb, marineb_state )
 
 	/* sound hardware */
 	MCFG_SPEAKER_STANDARD_MONO("mono")
-	MCFG_SOUND_ADD("ay1", AY8910, 1500000)
+	MCFG_SOUND_ADD("ay1", AY8910, SOUND_CLOCK)
 	MCFG_SOUND_ROUTE(ALL_OUTPUTS, "mono", 0.50)
 MACHINE_CONFIG_END
 
@@ -606,10 +606,10 @@ static MACHINE_CONFIG_DERIVED( wanted, marineb )
 	MCFG_SCREEN_UPDATE_DRIVER(marineb_state, screen_update_springer)
 
 	// sound hardware (PSG type verified only for bcruzm12)
-	MCFG_SOUND_REPLACE("ay1", AY8912, 1500000)
+	MCFG_SOUND_REPLACE("ay1", AY8912, SOUND_CLOCK)
 	MCFG_SOUND_ROUTE(ALL_OUTPUTS, "mono", 0.25)
 
-	MCFG_SOUND_ADD("ay2", AY8912, 1500000)
+	MCFG_SOUND_ADD("ay2", AY8912, SOUND_CLOCK)
 	MCFG_SOUND_ROUTE(ALL_OUTPUTS, "mono", 0.25)
 MACHINE_CONFIG_END
